Upgrade postgis container to postgres 12; use official postgis image
In process of upgrading from postgres 9 to 12, this commit updates
the local image to use postgres 12.  The `mdillon/postgis` image is
archived, and only supports up to postgres base version 11, so this
commit also switches to use the new official postgis image.

I was able to run "make docker/db" successfully, but I don't think
that actually hooks up the existing data in the database.  Need to install pipenv
and figure out how to load the local copy in to confirm the
existing data will be compatible with postgres 12.