Automate setting 8th grader age for cohorts
This is using a very Django template way to set the age of the 8th
graders from the first and latest cohorts. We can probably extract some
of the logic into some sort of filter or tag function but I think this
works for now.

We first get the current year with {% now 'Y' %} as a string. Then we
use the built-in add template tag to coerce the string into an integer
with current_year_str|add:"0". Then, we subtract the first cohort and
latest cohort years from the current year to get the number of years a
person has aged since they were in 8th grade. Finally, we add 13 to the
difference to get the person's age. Note that we are loading stats_tags
to get the subtract template filter.